fixing startup code in docker restserver

+++ rest_server/src/smfr.py
@@ -8,13 +8,15 @@ import shutil
 import click
 from flask_migrate import upgrade
 
+import utils
 
-current_dir = os.path.dirname(__file__)
-client_src = os.path.join(current_dir, '../../client')
-dst = os.path.join(current_dir, 'client/')
-if os.path.exists(dst):
-    shutil.rmtree(dst)
-shutil.copytree(client_src, dst)
+if not utils.running_in_docker():
+    current_dir = os.path.dirname(__file__)
+    client_src = os.path.join(current_dir, '../../client')
+    dst = os.path.join(current_dir, 'client/')
+    if os.path.exists(dst):
+        shutil.rmtree(dst)
+    shutil.copytree(client_src, dst)
 
 from client.api_client import ApiLocalClient
 from start import app
